+++ config/interpolate_funcs_test.go
@@ -6,6 +6,43 @@ import (
 	"testing"
 )
 
+func TestInterpolateFuncConcat(t *testing.T) {
+	cases := []struct {
+		Args   []string
+		Result string
+		Error  bool
+	}{
+		{
+			[]string{"foo", "bar", "baz"},
+			"foobarbaz",
+			false,
+		},
+
+		{
+			[]string{"foo", "bar"},
+			"foobar",
+			false,
+		},
+
+		{
+			[]string{"foo"},
+			"foo",
+			false,
+		},
+	}
+
+	for i, tc := range cases {
+		actual, err := interpolationFuncConcat(nil, tc.Args...)
+		if (err != nil) != tc.Error {
+			t.Fatalf("%d: err: %s", i, err)
+		}
+
+		if actual != tc.Result {
+			t.Fatalf("%d: bad: %#v", i, actual)
+		}
+	}
+}
+
 func TestInterpolateFuncFile(t *testing.T) {
 	tf, err := ioutil.TempFile("", "tf")
 	if err != nil {